shops/visualize_cnn_training/requirements.txt
tqdm
numpy
scipy
pandas
Pillow
tensorflow
umap-learn
streamlit
matplotlib
plotly


